requirements.txt
tinydb
watchdog
python-magic
rapidfuzz
requests
python-dotenv
colorlog

